requirements.txt
absl-py==1.2.0
adal==1.2.7
aiohttp==3.8.3
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.12
annotated-types==0.7.0
anyio==3.6.1
appdirs==1.4.4
applicationinsights==0.11.10
appnope==0.1.3
argcomplete==3.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
aspose-ocr-python-net==24.3.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
auto-sklearn==0.15.0
autokeras==2.0.0
azure-common==1.1.28
azure-core==1.30.1
azure-graphrbac==0.61.1
azure-identity==1.16.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.1.0
azureml==0.2.7
azureml-automl-core==1.56.0.post1
azureml-core==1.56.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.2
azureml-dataset-runtime==1.56.0
azureml-pipeline==1.56.0
azureml-pipeline-core==1.56.0
azureml-pipeline-steps==1.56.0
azureml-sdk==1.56.0
azureml-telemetry==1.56.0
azureml-train-automl-client==1.56.0
azureml-train-core==1.56.0
azureml-train-restclients-hyperdrive==1.56.0
Babel==2.10.3
backcall==0.2.0
backoff==1.11.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.3
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
branca==0.7.1
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.2.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.7
clikit==0.6.2
cloudpickle==2.2.1
cmdstanpy==1.0.8
ConfigSpace==0.4.21
contextlib2==21.6.0
convertdate==2.4.0
crashtest==0.3.1
cryptography==42.0.7
cycler==0.11.0
Cython==0.29.35
dask==2024.5.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distributed==2024.5.1
distro==1.9.0
Django==4.1.1
djangorestframework==3.14.0
dm-tree==0.1.8
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
easyocr==1.7.1
editdistance==0.8.1
efficientnet==1.0.0
einops==0.8.0
emcee==3.1.6
entmax==1.3
entrypoints==0.4
ephem==4.1.3
essential-generators==1.0
et-xmlfile==1.1.0
etils==1.7.0
eval_type_backport==0.2.0
executing==1.0.0
fastjsonschema==2.16.1
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.37.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.3.0
fusepy==3.0.1
future==1.0.0
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
gevent==22.10.2
google-auth==2.22.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.56.2
h11==0.12.0
h2==4.1.0
h2o==3.46.0.2
h5py==3.11.0
hijri-converter==2.2.4
holidays==0.17.2
hpack==4.0.0
httpcore==0.13.7
httpstan==4.9.0
httpx==0.19.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.3
imageio==2.31.1
imbalanced-learn==0.12.2
imblearn==0.0
imgaug==0.4.0
immutabledict==4.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
install==1.3.5
ipykernel==6.15.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.0
json5==0.9.10
jsonpickle==3.0.4
jsonpointer==2.3
jsonref==1.0.0.post1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions @ https://github.com/ipython-contrib/jupyter_contrib_nbextensions/tarball/master
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-lsp==1.5.1
jupyter-nbextensions-configurator==0.5.0
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyter_core==5.7.2
jupyterlab==3.4.6
jupyterlab-lsp==3.10.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
kaggle==1.5.13
kagglehub==0.2.5
keras==3.4.1
Keras-Applications==1.0.8
keras-core==0.1.7
keras-cv==0.9.0
keras-nlp==0.12.1
keras-ocr==0.9.3
Keras-Preprocessing==1.1.2
keras-tuner==1.4.7
kiwisolver==1.4.4
knack==0.11.0
korean-lunar-calendar==0.3.1
kt-legacy==1.0.5
lazy_loader==0.3
liac-arff==2.5.0
libclang==14.0.1
locket==1.0.0
LunarCalendar==0.0.9
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.3.2
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.4
msrest==0.7.1
msrestazure==0.6.4
mss==9.0.1
multidict==6.0.2
munch==4.0.0
Naked==0.1.32
namex==0.0.8
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.16.3
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
nodejs==0.1.1
norbert==0.2.1
notebook==6.4.12
notebook-shim==0.1.0
npm==0.1.1
numpy==1.24.4
oauthlib==3.2.0
openai==0.27.0
opencv-python==4.8.0.74
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.11.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
partd==1.4.2
pastel==0.2.1
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2023.12.1
pipreqs==0.5.0
pix2tex==0.1.2
pkginfo==1.10.0
platformdirs==4.2.0
portalocker==2.8.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
prophet==1.1.1
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.18.0
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.13.0
PyJWT==2.6.0
pylev==1.4.0
PyMeeus==0.5.11
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynisher==0.6.4
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.0.9
pyperclip==1.9.0
pyproject==1.3.1
PyRect==0.2.0
pyrfr==0.8.3
pyrsistent==0.18.1
PyScreeze==0.1.30
PySide6==6.4.0.1
PySide6-Addons==6.4.0.1
PySide6-Essentials==6.4.0.1
pysimdjson==5.0.2
PySocks==1.7.1
pystan==3.6.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytweening==1.2.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
redis==4.4.2
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3987==1.3.8
rich==13.7.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.0
setuptools-git==1.2
shapely==2.0.3
shellescape==3.8.1
shiboken6==6.4.0.1
simple_parsing==0.1.5
simplejson==3.17.6
six==1.16.0
smac==1.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spleeter==2.4.0
spotify==0.10.2
spotipy==2.22.1
sqlparse==0.4.3
stack-data==0.5.0
swagger-spec-validator==3.0.2
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.2
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.15.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.7.4
timm==0.5.4
tinycss2==1.1.1
tkcap==0.0.3
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.43.4
typer==0.3.2
typing_extensions==4.10.0
upbit==1.0.0
upbit-client==1.3.2.0
uri-template==1.2.0
urllib3==1.26.18
validators==0.28.1
virtualenv==20.25.3
wcwidth==0.2.5
webargs==8.2.0
webcolors==1.12
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==3.0.2
widgetsnbextension==4.0.3
wordcloud==1.9.2
wrapt==1.14.1
x-transformers==0.15.0
xgboost==2.0.3
xlrd==2.0.1
xyzservices==2024.4.0
yarg==0.1.9
yarl==1.8.1
zict==3.0.0
zipp==3.19.0
zope.event==4.5.0
zope.interface==5.5.2
